refactor: rename action package to 'context'
