+++ src/main/java/io/github/somesourcecode/someguiapi/scene/gui/ChestGui.java
@@ -24,7 +24,7 @@
 package io.github.somesourcecode.someguiapi.scene.gui;
 
 import io.github.somesourcecode.someguiapi.scene.*;
-import io.github.somesourcecode.someguiapi.scene.action.RenderContext;
+import io.github.somesourcecode.someguiapi.scene.context.RenderContext;
 import net.kyori.adventure.text.Component;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.HumanEntity;
