+++ src/main/java/io/github/somesourcecode/someguiapi/scene/lore/ReloadableParagraph.java
@@ -1,6 +1,6 @@
 package io.github.somesourcecode.someguiapi.scene.lore;
 
-import io.github.somesourcecode.someguiapi.scene.action.RenderContext;
+import io.github.somesourcecode.someguiapi.scene.context.RenderContext;
 
 /**
  * Represents a paragraph whose content depends on the context in which it is rendered.
